Hammett's equilibrium ExB shear model added.  There is a parallelization bug in this version that was found by Loureiro and Hammett, just before APS.  Do not use this version for parallel runs.
ingen updated.

Some post-processing codes and the like moved to Aux sub-directory.

This version was src_exb for several months on bassi.nersc.gov.  

Equilibrium ExB shear is an important new capability for the code, so
this is Version 5.0.  Will be tagged for release as 5.x when I figure
out what x is (incorporating some bug fixes and other upgrades from
the winter and spring of 2006/07).