Several changes in this update (MAB) :
1) Collision-operator error diagnostic modified to take 5-pt derivatives 
(instead of N-pt derivatives) when using Lagrangian scheme.  This was
 necessary to fix stability issues for runs with N large.

2) Amount and type of data written to files ".vres" (v-space resolution 
diagnostics) and ".cres" (collision-operator error diagnostic) changed 
to reduce file-size.

3) V-space resolution diagnostic updated to include errors in calculation of apar.

4) Vinor bug in v-space res. diag. fixed.  The bug was only present when using 
an odd number of energy grid points.

5) Bug in legendre transforms used for spectral v-space diagnostic fixed.

6) Scheme for initializing vnmult modified to fix bug.  Old scheme attempted
to initialize vnmult to value from restart file regardless of whether or 
not "check_restart" was chosen for delt_option, which led to file I/O errors.