modified init_lorentz_conserve, init_diffuse_conserve, conserve_lorentz, and conserve_diffuse to make conservation of number, momentum, and energy numerically exact, regardless of resolution.  this also required changes to the definitions of some of the v-dependent collision frequencies and the creation of a new array, vpdiff (and corresponding subroutine init_vpdiff).
this coding has been tested in normal operating mode to test that everything meant to be 
conserved is conserved when conservative=.true. (did not set this as the default though).
also seem to get correct (not exactly conserved) behavior when conservative=.false.
i did include coding for USE_L2E AND USE_LE, but i did not test this coding, as it does
not appear to be in use.