CMR, 17/7/2009:
Have set omprimfac=1 by default, and adjusted the normalisation of the 
omprimfac linear drive source term so that this provides the correct drive.

NB It is potentially copnfusing that in previous revisions of this branch the 
correct linear drive was obtained with omprimfac=2, and the default was 
omprimfac=0. Now omprimfac should be 1, which corresponds to the default!!