Modifications in this revision by GGH:
1) Solves a problem with the antenna using the same random number sequences:
a) Random number seeds are saved and written to restart runname.nc.### files
b) Note that the integer random number seeds are saved as floats in NetCDF file
   (I had problems getting the integer to save with the correct precision)
c) Usage: In input namelist &driver/:
   1) randomize=T specifies that random number genertor is initialized 
        based on the date and time (init_seed is still written)
   2) seed1 = 987654321 specifies first random number seed
   3) seed2 = 123456789 specifies secong random number seed
   4) Initial and final random number seeds for a run are save in the restart
       file  as init_seed(1:nseed) and fin_seed(1:nseed)
   5) The length of seed vector is the netcdf dimension nseed.
d) Note that the vector length of the random number seed for intrinsic random
   number function is implementation dependent.  It is assumed to be 2, but 
   could vary with other compilers.

2) The default behavior of USE_HDF5 in the Makefile is changed from 
"on" to "" (undefined) 

3) Bug Fix: Restart timestep input
a) Somewhere between r1242 and r1714, the order of initializations in
init_fields (fields.f90) was switched:
    i) In r1242 init_init_g was called before init_run_parameters
    ii) In r1714,  init_run_parameters was called before init_init_g 
b) Within init_run_parameters, the routine init_dt reads the old timestep
from the restart runname.nc.### files, but the name of these files is
initialized in init_init_g, so it must be called first.
c) I have switched the order back, but am wary of potential problems this 
may cause.