Helios login nodes use a different processor to the backend nodes so added additional compiler flag to target the backend processors. This gives about a 10% speed up in preliminary test but comes at the expense of the generated binaries no longer being executable on the login nodes.
For this reason a second makefile has been added which will generate code which is executable on both the front and back nodes at the expense of slower execution.
    => To use this pass the option GK_SYSTEM=helios_fe to make

An additional compiler flag has been added to both makefiles in order to disable a common remark about the string formats in use as it can obscure important warnings.