Testing some changes to the nonlinear/explicit term treatment. 1) Removing flow mode, not used (and not allowed to be anything other than off). This requires removing flow_mode from various test input files. 2) Adding explicit_schemes module, currently just provide a derived type to describe a generic multistep scheme. This is used in the new nonlinear treatment to define the Adams-Bashforth scheme. 3) Allow runtime selection of the explicit scheme order, to aid this gexp_1, gexp_2 etc. have been combined into a single gexp array and a small routine has been written to calculate the source given gexp. This can allow memory usage to be reduced (going to lower order) or errors to be reduced (going to higher order).
The code appears to work, though the current state includes a lot of testing code which can be removed in future. Currently a simple estimate of the error (based on comparing the source term calculated using the requested order and one order less) is calculated and reported to screen once every timestep. 

It should be relatively easy to either adapt the order of the scheme or the timestep to meet certain constraints on the error should this be desired.

Currently the maximum order supported is 5.