Adding 'store_history' logical flag to nonlinear_terms_knobs namelist. If true (not default) then the explicit source term, along with the current scheme order, will be saved to the restart file. This is then used when restoring from restart to restore the explicit source state. This makes restarted nonlinear runs follow more closely the trajectory of continued runs (e.g. two 600 step runs compared to a single 1200 step run should now agree better), provided 'force_maxwell_reinit=.false.' (not the default). Note there is a ~0.5*delt difference in the time values of restarted runs when compared to continued runs. Additionally, whilst the restarted runs track the continued run much closer than previously there is still some divergence after a finite number of steps.
Note that when changing the time step it is important to discard the existing explicit source history (as the scheme assumes constant time step), so this is now forced in gs2_reinit before the restart is saved.

Due to circular dependencies it was necessary to move some of the explicit source parameters to a small *_arrays.f90 type module.