Fix to subroutine exb_shear (in dist_fn.fpp) for new zeroed boundary regions during jumps, in box [or zero shat], linked case: zero the endpoint of the adjoining linked domain, to maintain single-valuedness. Tested linearly, ES with g_exb < 0 (checked phi), EM with g_exb > 0 (checked phi, apar and bpar). Two nonlinear tests (ES with g_exb < 0) seem to behave consistently with pre-patched code; not checked in detail.
Prior to this patch, double-values were introduced and led to a jump in both the field and its theta derivative across the
link between domains, which was persistent in time including across further jumps but eliminated at restart (with force_maxwell_reinit = T).

An outstanding issue is the effect of the physical discontinuity introduced where the new zeroed region joins the old boundary endpoint.
It may be better to propagate the old boundary endpoint values across the new region (hence also including the new endpoint that
is now the bounday), rather than zeroing. In general this would require communication. There is a similar issue at the other end,
where domains disappear and a previously internal endpoint suddenly becomes the new boundary. In the above linear tests, these
discontinuities led to oscillations in phi vs. theta (NB also with the pre-patched code), but smoothed out after only a few time steps.