Adding an example user routine used in selecting eigenvalues.
Rather than using EPSSetEigenvalueComparison, as originally intended, it has been necessary to use EPSSetArbitrarySelection for now. I believe this is due to a bug in slepc versions present at least up to v3.4.4. This bug appears to have been fixed in more recent versions, but as 3.4 is still quite common I'll leave the robust method in for now.