Merging some subcommunicator destruction code from branch eCSEwork in order to attempt to fix bug [#40]
Appears to work *but* re-initialisation steps now take much longer than in previous revisions. Results in having to destroy and reconstruct sub-communicators which should not change unless the data decomposition is changing or the processor layout is changed, which I don't think we ever do currently (possibly useful for trinity style runs). An alternative fix, which would lead to faster re-inits but without supporting a change in data decomposition, would be to ensure the new init level based approach to changing the time step uses the reset_fields routine rather than the finish_fields routine. This would be done in addition to this patch, which ensures the finish_fields_local routine does a complete tidy up.